+++ src/goryachev/swing/CheckForUpdate.java
@@ -5,7 +5,6 @@ import goryachev.common.log.Log;
 import goryachev.common.util.DotSeparatedVersion;
 import goryachev.common.util.UserException;
 import goryachev.i18n.TXT;
-import java.io.FileNotFoundException;
 import java.net.URL;
 
 
@@ -33,11 +32,11 @@ public class CheckForUpdate
 		{
 			siteVersion = reader.readString(url);
 		}
-		catch(FileNotFoundException e)
+		catch(Throwable e)
 		{
 			log.error(e);
 			URL u = new URL(url);
-			throw new UserException(TXT.get("CheckForUpdate.err.server problem", "Sorry, there seems to be a problem contacting server [{0}]", u.getHost()));
+			throw new UserException(TXT.get("CheckForUpdate.err.server problem", "Sorry, there seems to be a problem contacting server\n[{0}]", u.getHost()));
 		}
 	}