+++ Release Notes.md
@@ -3,6 +3,7 @@
 
 ## TBD
 - create parent folder(s) when creating a new data file
+- fixed error handling in check for update
 
 ## 2022.0124.1555
 - added generate password dialog